// argon-dashboard-angular-master/src/app/models/user.ts
export class User {
  id: number;
  username: string;
  password: string;
  name: string;
  email: string;
  gender: boolean;
  maritalStatus: boolean;
  birthday: string;
  nationality: number;
  author: string;
  positionName: string;
  departmentName: string;
  managerName: string;
  hireDate: string;
}